--- lib/agala_email_responser.ex
@@ -1,7 +1,0 @@
-defmodule Agala.Provider.Email.Responser do
-  @moduledoc """
-  Module providing email connection send
-  """
-  use Agala.Bot.Responser
-  alias Agala.BotParams
-end
